requirements.txt
whatsapp-chatbot-python
sqlalchemy
alembic
pytest
python-dotenv
psycopg[binary]
